ai-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
numpy>=1.24.0
trimesh>=4.0.0
rembg>=2.0.0
einops>=0.7.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.25.0
safetensors>=0.4.0
omegaconf>=2.3.0
replicate>=0.25.0
requests>=2.31.0
